@@ -10,7 +10,7 @@ if (project.hasProperty("sma2.signing") &&
         rootProject.file("apikey.properties").exists()) {
 
     android {
-        compileSdkVersion 26
+        compileSdkVersion 28
         lintOptions {
             abortOnError false
             checkReleaseBuilds false
@@ -18,7 +18,7 @@ if (project.hasProperty("sma2.signing") &&
         defaultConfig {
             applicationId "com.sma2.sma2"
             minSdkVersion 21
-            targetSdkVersion 26
+            targetSdkVersion 28
             versionCode 1
             versionName "1.0"
             testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
@@ -56,7 +56,7 @@ if (project.hasProperty("sma2.signing") &&
     }
 } else {
     android {
-        compileSdkVersion 26
+        compileSdkVersion 28
         lintOptions {
             abortOnError false
             checkReleaseBuilds false
@@ -64,7 +64,7 @@ if (project.hasProperty("sma2.signing") &&
         defaultConfig {
             applicationId "com.sma2.sma2"
             minSdkVersion 21
-            targetSdkVersion 26
+            targetSdkVersion 28
             versionCode 1
             versionName "1.0"
             testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
@@ -89,35 +89,34 @@ greendao{
 
 dependencies {
     implementation fileTree(include: ['*.jar'], dir: 'libs')
-    implementation 'com.android.support:appcompat-v7:26.1.0'
-    implementation 'com.android.support.constraint:constraint-layout:1.1.2'
-    implementation 'com.android.support:support-v13:26.1.0'
+    implementation 'androidx.appcompat:appcompat:1.0.0'
+    implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
+    implementation 'androidx.legacy:legacy-support-v13:1.0.0'
     implementation 'com.opencsv:opencsv:4.3.1'
     implementation 'com.jjoe64:graphview:4.2.1'
-    implementation 'com.android.support:support-v4:26.1.0'
-    implementation 'com.android.support:recyclerview-v7:26.1.0'
+    implementation 'androidx.recyclerview:recyclerview:1.0.0'
     implementation 'org.apache.commons:commons-math3:3.6.1'
     implementation 'org.apache.commons:commons-lang3:3.6'
     implementation group: 'uk.me.berndporr', name:'iirj', version: '1.1'
 
     testImplementation 'junit:junit:4.12'
-    androidTestImplementation 'com.android.support.test:runner:1.0.2'
-    androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
+    androidTestImplementation 'androidx.test.ext:junit:1.1.1'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.1.0'
     implementation 'org.greenrobot:greendao:3.2.2' // add library
     implementation 'com.squareup.okhttp3:okhttp:3.11.0'
     implementation 'com.alibaba:fastjson:1.1.70.android'
     implementation 'us.feras.mdv:markdownview:1.1.0'
-    implementation 'com.android.support:cardview-v7:26.1.0'
+    implementation 'androidx.cardview:cardview:1.0.0'
     testImplementation 'junit:junit:4.12'
     implementation 'com.github.PhilJay:MPAndroidChart:v3.1.0'
 
-    testImplementation 'com.android.support.test:runner:1.0.2'
-    testImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
+    testImplementation 'androidx.test.ext:junit:1.1.1'
+    testImplementation 'androidx.test.espresso:espresso-core:3.1.0'
     testImplementation 'org.mockito:mockito-core:1.10.19'
     testImplementation 'org.robolectric:robolectric:3.4.2'
-    implementation 'com.android.support:design:26.1.0'
+    implementation 'com.google.android.material:material:1.0.0'
 
-    implementation 'com.android.support:support-annotations:27.1.1'
+    implementation 'androidx.annotation:annotation:1.1.0'
     implementation 'com.mcxiaoke.volley:library:1.0.19'
     implementation 'com.google.firebase:firebase-core:16.0.9'
     implementation 'com.crashlytics.sdk.android:crashlytics:2.10.1'
