@@ -5,9 +5,8 @@ import android.content.Context;
 import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabase;
 import android.database.sqlite.SQLiteOpenHelper;
-import android.support.annotation.Nullable;
-import android.util.Log;
-import android.widget.Toast;
+import androidx.annotation.Nullable;
+
 
 import java.util.ArrayList;
 import java.util.List;
