@@ -1,11 +1,12 @@
 package com.sma2.sma2.ExerciseFragments;
 
-import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.CheckBox;
 import android.widget.TextView;
+
+import androidx.recyclerview.widget.RecyclerView;
 
 import com.sma2.sma2.DataAccess.ScheduledExercise;
 import com.sma2.sma2.R;
