@@ -4,9 +4,7 @@ import android.app.Activity;
 import android.content.Context;
 import android.content.Intent;
 import android.media.MediaPlayer;
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentActivity;
-import android.support.v4.app.FragmentTransaction;
+
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -16,10 +14,7 @@ import android.widget.TextView;
 
 import com.sma2.sma2.DataAccess.Exercise;
 import com.sma2.sma2.DataAccess.ScheduledExercise;
-import com.sma2.sma2.ExerciseFragments.ExerciseFragment;
-import com.sma2.sma2.ExerciseFragments.ExerciseInstructions;
 import com.sma2.sma2.ExercisesActivity;
-import com.sma2.sma2.MainActivityMenu;
 import com.sma2.sma2.R;
 
 import java.util.ArrayList;
