@@ -2,10 +2,11 @@ package com.sma2.sma2.ExerciseList;
 
 import android.content.Context;
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import androidx.fragment.app.Fragment;
 
 import com.sma2.sma2.R;
 
