@@ -3,12 +3,14 @@ package com.sma2.sma2;
 import android.content.Context;
 import android.net.Uri;
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
 import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ListView;
+
+import androidx.fragment.app.Fragment;
+
 import com.sma2.sma2.DataAccess.ExerciseSessionManager;
 
 import com.sma2.sma2.DataAccess.Exercise;
