@@ -3,27 +3,24 @@ package com.sma2.sma2;
 import android.content.Context;
 import android.content.Intent;
 import android.content.SharedPreferences;
-import android.graphics.Point;
 import android.net.Uri;
 import android.os.Bundle;
 import android.preference.PreferenceManager;
-import android.support.v4.app.Fragment;
-import android.view.Display;
+
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
 import android.widget.ImageButton;
 import android.widget.ImageView;
-import android.widget.LinearLayout;
 import android.widget.ProgressBar;
 import android.widget.TextView;
-import android.widget.Toast;
+
+import androidx.fragment.app.Fragment;
 
 import com.github.mikephil.charting.charts.RadarChart;
 import com.sma2.sma2.DataAccess.FeatureDA;
 import com.sma2.sma2.DataAccess.FeatureDataService;
-import com.sma2.sma2.FeatureExtraction.Speech.features.RadarFeatures;
 
 import java.util.Calendar;
 import java.util.Date;
